Refactor cache to separate module
